new(vx-demo): convert LinkTypes to codesandbox
