+++ packages/vx-demo/src/docs-v2/examples/vx-linktypes/Example.tsx
@@ -1,0 +1,208 @@
+import React, { useState } from 'react';
+import { Group } from '@vx/group';
+import { hierarchy, Tree } from '@vx/hierarchy';
+import { LinearGradient } from '@vx/gradient';
+import { pointRadial } from 'd3-shape';
+import useForceUpdate from './useForceUpdate';
+import LinkControls from './LinkControls';
+import getLinkComponent from './getLinkComponent';
+
+interface TreeNode {
+  name: string;
+  isExpanded?: boolean;
+  children?: TreeNode[];
+}
+
+const data: TreeNode = {
+  name: 'T',
+  children: [
+    {
+      name: 'A',
+      children: [
+        { name: 'A1' },
+        { name: 'A2' },
+        { name: 'A3' },
+        {
+          name: 'C',
+          children: [
+            {
+              name: 'C1',
+            },
+            {
+              name: 'D',
+              children: [
+                {
+                  name: 'D1',
+                },
+                {
+                  name: 'D2',
+                },
+                {
+                  name: 'D3',
+                },
+              ],
+            },
+          ],
+        },
+      ],
+    },
+    { name: 'Z' },
+    {
+      name: 'B',
+      children: [{ name: 'B1' }, { name: 'B2' }, { name: 'B3' }],
+    },
+  ],
+};
+
+const defaultMargin = { top: 30, left: 30, right: 30, bottom: 30 };
+
+type Props = {
+  width: number;
+  height: number;
+  margin?: { top: number; right: number; bottom: number; left: number };
+};
+
+export default function Example({
+  width: totalWidth,
+  height: totalHeight,
+  margin = defaultMargin,
+}: Props) {
+  const [layout, setLayout] = useState<string>('cartesian');
+  const [orientation, setOrientation] = useState<string>('horizontal');
+  const [linkType, setLinkType] = useState<string>('diagonal');
+  const [stepPercent, setStepPercent] = useState<number>(0.5);
+  const forceUpdate = useForceUpdate();
+
+  const innerWidth = totalWidth - margin.left - margin.right;
+  const innerHeight = totalHeight - margin.top - margin.bottom;
+
+  let origin: { x: number; y: number };
+  let sizeWidth: number;
+  let sizeHeight: number;
+
+  if (layout === 'polar') {
+    origin = {
+      x: innerWidth / 2,
+      y: innerHeight / 2,
+    };
+    sizeWidth = 2 * Math.PI;
+    sizeHeight = Math.min(innerWidth, innerHeight) / 2;
+  } else {
+    origin = { x: 0, y: 0 };
+    if (orientation === 'vertical') {
+      sizeWidth = innerWidth;
+      sizeHeight = innerHeight;
+    } else {
+      sizeWidth = innerHeight;
+      sizeHeight = innerWidth;
+    }
+  }
+
+  const LinkComponent = getLinkComponent({ layout, linkType, orientation });
+
+  return totalWidth < 10 ? null : (
+    <div>
+      <LinkControls
+        layout={layout}
+        orientation={orientation}
+        linkType={linkType}
+        stepPercent={stepPercent}
+        setLayout={setLayout}
+        setOrientation={setOrientation}
+        setLinkType={setLinkType}
+        setStepPercent={setStepPercent}
+      />
+      <svg width={totalWidth} height={totalHeight}>
+        <LinearGradient id="links-gradient" from="#fd9b93" to="#fe6e9e" />
+        <rect width={totalWidth} height={totalHeight} rx={14} fill="#272b4d" />
+        <Group top={margin.top} left={margin.left}>
+          <Tree
+            root={hierarchy(data, d => (d.isExpanded ? null : d.children))}
+            size={[sizeWidth, sizeHeight]}
+            separation={(a, b) => (a.parent === b.parent ? 1 : 0.5) / a.depth}
+          >
+            {tree => (
+              <Group top={origin.y} left={origin.x}>
+                {tree.links().map((link, i) => (
+                  <LinkComponent
+                    key={i}
+                    data={link}
+                    percent={stepPercent}
+                    stroke="white"
+                    strokeOpacity={0.3}
+                    strokeWidth="1"
+                    fill="none"
+                  />
+                ))}
+
+                {tree.descendants().map((node, key) => {
+                  const width = 40;
+                  const height = 20;
+
+                  let top: number;
+                  let left: number;
+                  if (layout === 'polar') {
+                    const [radialX, radialY] = pointRadial(node.x, node.y);
+                    top = radialY;
+                    left = radialX;
+                  } else if (orientation === 'vertical') {
+                    top = node.y;
+                    left = node.x;
+                  } else {
+                    top = node.x;
+                    left = node.y;
+                  }
+
+                  return (
+                    <Group top={top} left={left} key={key}>
+                      {node.depth === 0 && (
+                        <circle
+                          r={12}
+                          fill="url('#links-gradient')"
+                          onClick={() => {
+                            node.data.isExpanded = !node.data.isExpanded;
+                            console.log(node);
+                            forceUpdate();
+                          }}
+                        />
+                      )}
+                      {node.depth !== 0 && (
+                        <rect
+                          height={height}
+                          width={width}
+                          y={-height / 2}
+                          x={-width / 2}
+                          fill="#272b4d"
+                          stroke={node.data.children ? '#03c0dc' : '#26deb0'}
+                          strokeWidth={1}
+                          strokeDasharray={node.data.children ? '0' : '2,2'}
+                          strokeOpacity={node.data.children ? 1 : 0.6}
+                          rx={node.data.children ? 0 : 10}
+                          onClick={() => {
+                            node.data.isExpanded = !node.data.isExpanded;
+                            console.log(node);
+                            forceUpdate();
+                          }}
+                        />
+                      )}
+                      <text
+                        dy=".33em"
+                        fontSize={9}
+                        fontFamily="Arial"
+                        textAnchor="middle"
+                        style={{ pointerEvents: 'none' }}
+                        fill={node.depth === 0 ? '#71248e' : node.children ? 'white' : '#26deb0'}
+                      >
+                        {node.data.name}
+                      </text>
+                    </Group>
+                  );
+                })}
+              </Group>
+            )}
+          </Tree>
+        </Group>
+      </svg>
+    </div>
+  );
+}
