+++ packages/vx-demo/src/docs-v2/examples/vx-linktypes/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@vx/demo-linktypes",
+  "description": "Standalone vx link types demo.",
+  "main": "index.tsx",
+  "dependencies": {
+    "@babel/runtime": "^7.8.4",
+    "@types/react": "^16",
+    "@types/react-dom": "^16",
+    "@vx/gradient": "latest",
+    "@vx/group": "latest",
+    "@vx/hierarchy": "latest",
+    "@vx/responsive": "latest",
+    "@vx/shape": "latest",
+    "d3-shape": "1.3.7",
+    "react": "^16.8",
+    "react-dom": "^16.8",
+    "react-scripts-ts": "3.1.0",
+    "typescript": "^3"
+  },
+  "keywords": [
+    "visualization",
+    "d3",
+    "react",
+    "vx",
+    "link",
+    "shape"
+  ]
+}
